# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareerCraft",
    "description": "Resume parsing, job matching and cover letter drafting \u2014 all local.\n\nThis API backs the CareerCraft web UI. The primary interface to the same\nfunctionality is the MCP server (`careercraft-mcp`), which plugs into Claude\nDesktop, Claude Code and Cursor.\n\n`GET /api/capabilities` reports which optional features this install has;\nanything unavailable comes with the command that enables it.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/jobs/search": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Search Jobs",
        "operationId": "search_jobs_api_jobs_search_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSearchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/recent": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Recent Jobs",
        "operationId": "recent_jobs_api_jobs_recent_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "title": "Response Recent Jobs Api Jobs Recent Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/saved": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Saved Jobs",
        "operationId": "saved_jobs_api_jobs_saved_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SavedJob"
                  },
                  "title": "Response Saved Jobs Api Jobs Saved Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/save": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Save Job",
        "operationId": "save_job_api_jobs__job_id__save_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveJobRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Unsave Job",
        "operationId": "unsave_job_api_jobs__job_id__save_delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resumes/upload": {
      "post": {
        "tags": [
          "resumes"
        ],
        "summary": "Upload Resume",
        "description": "Accept an upload and parse it.\n\nThe client-supplied filename is discarded rather than sanitised \u2014 only its\nextension survives \u2014 and the body is streamed to disk with the size cap\nenforced as bytes arrive, so an oversized upload cannot fill the disk\nbefore anyone checks.",
        "operationId": "upload_resume_api_resumes_upload_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume_api_resumes_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParsedResume"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resumes/parse": {
      "post": {
        "tags": [
          "resumes"
        ],
        "summary": "Parse Resume Text",
        "description": "Parse pasted resume text.\n\nThere is deliberately no ``path=`` equivalent here. Over HTTP the caller\nis not necessarily the person running the server, and a tool that reads\nserver-side paths on request is a file-disclosure primitive.",
        "operationId": "parse_resume_text_api_resumes_parse_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParseTextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParsedResume"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resumes": {
      "get": {
        "tags": [
          "resumes"
        ],
        "summary": "List Resumes",
        "operationId": "list_resumes_api_resumes_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumeSummary"
                  },
                  "title": "Response List Resumes Api Resumes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resumes/{resume_id}": {
      "get": {
        "tags": [
          "resumes"
        ],
        "summary": "Get Resume",
        "operationId": "get_resume_api_resumes__resume_id__get",
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resume Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParsedResume"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "resumes"
        ],
        "summary": "Delete Resume",
        "operationId": "delete_resume_api_resumes__resume_id__delete",
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resume Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/match": {
      "post": {
        "tags": [
          "matching"
        ],
        "summary": "Match Jobs",
        "operationId": "match_jobs_api_match_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/letters": {
      "post": {
        "tags": [
          "letters"
        ],
        "summary": "Generate Letter",
        "operationId": "generate_letter_api_letters_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LetterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LetterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/letters/{letter_id}": {
      "get": {
        "tags": [
          "letters"
        ],
        "summary": "Get Letter",
        "operationId": "get_letter_api_letters__letter_id__get",
        "parameters": [
          {
            "name": "letter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Letter Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverLetter"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_resume_api_resumes_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "A .pdf, .docx, .txt or .md resume."
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_resume_api_resumes_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Job": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Stable content hash; the same posting always gets the same id."
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company": {
            "type": "string",
            "title": "Company",
            "default": ""
          },
          "location": {
            "type": "string",
            "title": "Location",
            "default": "Remote"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "url": {
            "type": "string",
            "title": "Url",
            "default": ""
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": ""
          },
          "level": {
            "type": "string",
            "title": "Level",
            "default": ""
          },
          "job_type": {
            "type": "string",
            "title": "Job Type",
            "default": ""
          },
          "salary": {
            "type": "string",
            "title": "Salary",
            "default": ""
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "Job",
        "description": "A single job posting, normalised across sources."
      },
...
      "JobSearchResult": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/Job"
            },
            "type": "array",
            "title": "Jobs"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "query": {
            "type": "string",
            "title": "Query",
            "default": ""
          },
          "location": {
            "type": "string",
            "title": "Location",
            "default": ""
          },
          "from_cache": {
            "type": "boolean",
            "title": "From Cache",
            "default": false
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "jobs",
          "total"
        ],
        "title": "JobSearchResult"
      },
...
      "ParseTextRequest": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "persist": {
            "type": "boolean",
            "title": "Persist",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ParseTextRequest"
      },
...
      "ResumeSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "skills_count": {
            "type": "integer",
            "title": "Skills Count",
            "default": 0
          },
          "experience_count": {
            "type": "integer",
            "title": "Experience Count",
            "default": 0
          },
          "parsed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Parsed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "parsed_at"
        ],
        "title": "ResumeSummary",
        "description": "A row in ``list_resumes`` \u2014 enough to pick one, not the whole document."
      },
      "SaveJobRequest": {
        "properties": {
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SaveJobRequest"
      },
      "SavedJob": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/Job"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job"
        ],
        "title": "SavedJob"
      },
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "default": ""
          },
          "location": {
            "type": "string",
            "title": "Location",
            "default": ""
          },
          "limit": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 25
          },
          "remote_only": {
            "type": "boolean",
            "title": "Remote Only",
            "default": true
          },
          "refresh": {
            "type": "boolean",
            "title": "Refresh",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SearchRequest"
      },
...
    }
  }
}